chore: move credential helpers to its own file
